Add files via upload
This portion of the code will need to be added to all the the item pages for the cart to work. change the details to the item name and the price to the item price. This only shows up on the carts page after the item is added. the <a href="shortshome.html"> will get changed to the actual homepage of the website and we can add a line to return to the homepage for each item tip as well.